+++ src/main/java/bridge/dtos/GameResultDto.java
@@ -1,6 +1,6 @@
 package bridge.dtos;
 
-public record GameResult(MapState mapState, boolean isSuccess, int tryCount) {
+public record GameResultDto(MapStateDto mapState, boolean isSuccess, int tryCount) {
 
     public String getGameResult() {
         if (isSuccess) {
