+++ src/main/java/bridge/dtos/MapStateDto.java
@@ -3,7 +3,7 @@ package bridge.dtos;
 import bridge.domain.BridgeCell;
 import java.util.List;
 
-public record MapState(List<SingleMapState> results) {
+public record MapStateDto(List<SingleMapState> results) {
     public record SingleMapState(BridgeCell userAnswer, boolean isCorrect) {
 
     }
