+++ src/main/java/bridge/view/OutputView.java
@@ -1,9 +1,9 @@
 package bridge.view;
 
 import bridge.domain.BridgeCell;
-import bridge.dtos.GameResult;
-import bridge.dtos.MapState;
-import bridge.dtos.MapState.SingleMapState;
+import bridge.dtos.GameResultDto;
+import bridge.dtos.MapStateDto;
+import bridge.dtos.MapStateDto.SingleMapState;
 import java.util.List;
 
 /**
@@ -21,7 +21,7 @@ public class OutputView {
      * <p>
      * 출력을 위해 필요한 메서드의 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
      */
-    public void printMap(MapState MapState) {
+    public void printMap(MapStateDto MapState) {
         List<SingleMapState> results = MapState.results();
         printUpResult(results);
         printNewLine();
@@ -33,7 +33,7 @@ public class OutputView {
      * <p>
      * 출력을 위해 필요한 메서드의 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
      */
-    public void printResult(GameResult gameResult) {
+    public void printResult(GameResultDto gameResult) {
         System.out.println("최종 게임 결과");
         this.printMap(gameResult.mapState());
         printNewLine();
