+++ src/test/java/bridge/view/OutputViewTest.java
@@ -1,9 +1,9 @@
 package bridge.view;
 
 import bridge.domain.BridgeCell;
-import bridge.dtos.GameResult;
-import bridge.dtos.MapState;
-import bridge.dtos.MapState.SingleMapState;
+import bridge.dtos.GameResultDto;
+import bridge.dtos.MapStateDto;
+import bridge.dtos.MapStateDto.SingleMapState;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 
@@ -12,7 +12,7 @@ class OutputViewTest {
     @Test
     void 이동한_다리_상태_출력_테스트() {
         OutputView outputView = new OutputView();
-        MapState mapState = new MapState(
+        MapStateDto mapState = new MapStateDto(
                 List.of(
                         new SingleMapState(BridgeCell.UP, true),
                         new SingleMapState(BridgeCell.UP, true),
@@ -27,7 +27,7 @@ class OutputViewTest {
     @Test
     void 최종_결과_테스트() {
         OutputView outputView = new OutputView();
-        MapState mapState = new MapState(
+        MapStateDto mapState = new MapStateDto(
                 List.of(
                         new SingleMapState(BridgeCell.UP, true),
                         new SingleMapState(BridgeCell.UP, true),
@@ -35,7 +35,7 @@ class OutputViewTest {
                         new SingleMapState(BridgeCell.DOWN, false)
                 )
         );
-        GameResult gameResult = new GameResult(
+        GameResultDto gameResult = new GameResultDto(
                 mapState,
                 false,
                 2
